+++ helpers/b_reading-clear.sh
@@ -1,1 +1,0 @@
-ls: cannot access 'targets/ReadingList/sources': No such file or directory